// src/cloudapp/api-client/types.ts
type AxiosOrganization = {
  metadata: Record<string, any>;
  message: string;
  id: number;
  hashid: string;
  created_at: string;
  updated_at: string;
  api_key: null;
  billing_info: {
    attributes: {
      object: string;
      card_type: string;
      first_six: string;
      last_four: string;
      exp_month: number;
      exp_year: number;
    };
  };
  collections_count: number;
  custom_domain_name: string;
  custom_storage: boolean;
  features: {
    sso: boolean;
    security: boolean;
    custom_s3: boolean;
    custom_domain: boolean;
  };
  items_count: number;
  limits: {
    storage: number;
    bandwidth: number;
    gif_length: number;
    daily_uploads: number;
    max_upload_size: number;
    monthly_uploads: number;
    max_video_length: number;
  };
  name: string;
  options: null;
  personal: boolean;
  plan_name: string;
  preferences: {
    annotations_restricted: null;
    allow_anonymous_comments: null;
    allow_external_comments: null;
    allow_overwrite_item_expire: null;
    authentication: {
      basic: boolean;
      google: boolean;
    };
    comments_disabled: null;
    custom_domain: {};
    custom_storage: {};
    disallow_invite_by_members: null;
    download_button: null;
    hide_cloudapp_branding: boolean;
    ip_withelist_for_item: null;
    ip_withelist_for_account: null;
    ip_withelist_for_item_enabled: null;
    ip_withelist_for_account_enabled: null;
    items_expire_in: null;
    items_retention: null;
    items_remove_action: null;
    items_remove_enabled_at: null;
    items_sharing: 'un';
    items_password: null;
    logo: null;
    redirect_root_url: null;
    restrict_item_access: null;
    restrict_item_invitations: null;
    sharing: null;
    v1: null;
    v1_id: null;
    v1_custom_domain: null;
    watermark_url: null;
    webhook_url: null;
    limit_workspaces: false;
    sso_with_scim: false;
    discoverable: null;
  };
  owner_id: number;
  recurly_account_id: string;
  plans_url: string;
  plan_trial: boolean;
  plan_code: string;
  trial_expires_at: null;
};

export type AxiosAccount = {
  metadata: Record<string, any>;
  message: string;
  id: number;
  confirmed: boolean;
  email: string;
  plan_name: string;
  jwt: string;
  last_accessed_org_id: number;
  last_accessed_org_hashid: string;
  gif_recording_limit: number;
  recording_limit: number;
  dashboard_url: string;
  pusher_channel: string;
  organizations: AxiosOrganization[];
};

export type AxiosItem = {
  metadata: Record<string, any>;
  message: string;
  slug: string;
  name: string;
  created_at: string;
  updated_at: string;
  item_type: 'image' | 'video' | 'other' | 'archive' | 'file';
  has_security: boolean;
  view_counter: number;
  content_url: string;
  remote_url: string;
  source: string;
  thumbnail_url: string;
  share_url: string;
  id: string;
  favorite: boolean;
  collections: [];
  download_url: string;
  status: 'completed';
  security: {
    sharing: 'un';
    password: null;
    password_protected: boolean;
    expires: 'ne';
    expires_after_views: null;
    expires_at: null;
    expired: boolean;
    specific_users: [];
  };
  action_type: 'sc';
  checksum: boolean;
  file_name: boolean;
  type: 'image' | 'video' | 'other' | 'archive' | 'file';
  is_video: boolean;
  uploading: boolean;
  custom_domain_name: string;
  user_id: number;
  expires_at: null;
  expires_after_views: number;
  comments_disabled: null;
  upload_url: string;
  file_ext: string;
  m3u8: boolean;
  file_name_without_ext: string;
  public_visits_count: number;
  source_url: string;
  organization: AxiosOrganization;
};